+++ CHANGELOG.md
@@ -1,12 +1,19 @@
 # 更新
 
+## 2021-10-09
+
+* 更新topgg-autoposter，舊版時常CRASH，令HKTRPG出錯
+* 更新了PLURK，令速度提升
+* 改良CODE，移除了沒用的AWAIT ASYNC，令程式反應更快
+* 部份使用了IIFE
+
 ## 2021-09-30
 
 * 新增匯出團錄時，可以去掉不必要的日期標示，Choice 排序功能顯示改良，.CC7build random
 
 ## 2021-09-18
 
-* showMeAtTheWorld
+* 增加功能showMeAtTheWorld
 
 ## 2021-09-10